refactor(server): cleanup analyzer access
