+++ server/src/parser/common/types/index.ts
@@ -943,3 +943,5 @@ export const visibilityPrecedence: AstVisibility[] = [
   "external",
   "public",
 ];
+
+export type Overwrite<T, U> = Pick<T, Exclude<keyof T, keyof U>> & U;
