+++ server/src/services/completion/types.ts
@@ -1,0 +1,6 @@
+import { DocumentsAnalyzerMap } from "@common/types";
+
+export type ProjectContext = {
+  projectBasePath: string;
+  solFileIndex: DocumentsAnalyzerMap;
+};
