+++ server/src/services/definition/onDefinition.ts
@@ -1,38 +1,62 @@
-import { SolidityNavigation } from "@services/navigation/SolidityNavigation";
+import {
+  DocumentAnalyzer,
+  VSCodePosition,
+  Node,
+  Location,
+} from "@common/types";
+import { getParserPositionFromVSCodePosition, getRange } from "@common/utils";
+import { convertHardhatUriToVscodeUri } from "@utils/index";
+import { onCommand } from "@utils/onCommand";
 import { DefinitionParams } from "vscode-languageserver/node";
 import { ServerState } from "../../types";
-import { getUriFromDocument } from "../../utils/index";
 
 export const onDefinition = (serverState: ServerState) => {
   return (params: DefinitionParams) => {
-    const { logger } = serverState;
-
-    logger.trace("onDefinition");
-
     try {
-      const document = serverState.documents.get(params.textDocument.uri);
-
-      if (!document) {
-        return;
-      }
-
-      const documentURI = getUriFromDocument(document);
-      const documentAnalyzer =
-        serverState.analyzer.getDocumentAnalyzer(documentURI);
-
-      if (!documentAnalyzer.isAnalyzed) {
-        return;
-      }
-
-      return serverState.telemetry.trackTimingSync("onDefinition", () =>
-        new SolidityNavigation(serverState.analyzer).findDefinition(
-          documentURI,
-          params.position,
-          documentAnalyzer.analyzerTree.tree
-        )
+      return onCommand(
+        serverState,
+        "onDefinition",
+        params.textDocument.uri,
+        (documentAnalyzer) =>
+          findDefinition(serverState, documentAnalyzer, params.position)
       );
     } catch (err) {
-      logger.error(err);
+      serverState.logger.error(err);
     }
   };
 };
+
+function findDefinition(
+  serverState: ServerState,
+  documentAnalyzer: DocumentAnalyzer,
+  position: VSCodePosition
+) {
+  const definitionNode = documentAnalyzer.searcher.findDefinitionNodeByPosition(
+    documentAnalyzer.uri,
+    getParserPositionFromVSCodePosition(position),
+    documentAnalyzer.analyzerTree.tree
+  );
+
+  if (!definitionNode) {
+    return undefined;
+  }
+
+  const location = resolveLocationFrom(definitionNode);
+
+  if (!location) {
+    return undefined;
+  }
+
+  return {
+    uri: convertHardhatUriToVscodeUri(definitionNode.uri),
+    range: getRange(location),
+  };
+}
+
+function resolveLocationFrom(definitionNode: Node): Location | undefined {
+  if (definitionNode.type === "ImportDirective") {
+    return definitionNode.astNode.loc;
+  }
+
+  return definitionNode.nameLoc ?? definitionNode.astNode.loc;
+}
