+++ server/src/services/typeDefinition/onTypeDefinition.ts
@@ -1,38 +1,57 @@
-import { SolidityNavigation } from "@services/navigation/SolidityNavigation";
+import {
+  DocumentAnalyzer,
+  VSCodePosition,
+  Node,
+  Location,
+  Overwrite,
+} from "@common/types";
+import { getParserPositionFromVSCodePosition, getRange } from "@common/utils";
+import { convertHardhatUriToVscodeUri } from "@utils/index";
+import { onCommand } from "@utils/onCommand";
 import { TypeDefinitionParams } from "vscode-languageserver/node";
 import { ServerState } from "../../types";
-import { getUriFromDocument } from "../../utils/index";
 
 export const onTypeDefinition = (serverState: ServerState) => {
   return (params: TypeDefinitionParams) => {
-    const { logger } = serverState;
-
-    logger.trace("onTypeDefinition");
-
     try {
-      const document = serverState.documents.get(params.textDocument.uri);
-
-      if (!document) {
-        return;
-      }
-
-      const documentURI = getUriFromDocument(document);
-      const documentAnalyzer =
-        serverState.analyzer.getDocumentAnalyzer(documentURI);
-
-      if (!documentAnalyzer.isAnalyzed) {
-        return;
-      }
-
-      return serverState.telemetry.trackTimingSync("onTypeDefinition", () =>
-        new SolidityNavigation(serverState.analyzer).findTypeDefinition(
-          documentURI,
-          params.position,
-          documentAnalyzer.analyzerTree.tree
-        )
+      return onCommand(
+        serverState,
+        "onTypeDefinition",
+        params.textDocument.uri,
+        (documentAnalyzer) =>
+          findTypeDefinition(serverState, documentAnalyzer, params.position)
       );
     } catch (err) {
-      logger.error(err);
+      serverState.logger.error(err);
+
+      return null;
     }
   };
 };
+
+function findTypeDefinition(
+  serverState: ServerState,
+  documentAnalyzer: DocumentAnalyzer,
+  position: VSCodePosition
+) {
+  const definitionNode = documentAnalyzer.searcher.findDefinitionNodeByPosition(
+    documentAnalyzer.uri,
+    getParserPositionFromVSCodePosition(position),
+    documentAnalyzer.analyzerTree.tree
+  );
+
+  if (!definitionNode) {
+    return [];
+  }
+
+  return definitionNode
+    .getTypeNodes()
+    .filter(
+      (typeNode): typeNode is Overwrite<Node, { nameLoc: Location }> =>
+        typeNode.nameLoc !== undefined
+    )
+    .map((typeNode) => ({
+      uri: convertHardhatUriToVscodeUri(typeNode.uri),
+      range: getRange(typeNode.nameLoc),
+    }));
+}
