+++ server/test/setup.ts
@@ -5,6 +5,7 @@ const aliases = {
   "@analyzer": __dirname + "/../src/parser/analyzer",
   "@common": __dirname + "/../src/parser/common",
   "@services": __dirname + "/../src/services",
+  "@utils": __dirname + "/../src/utils",
 };
 
 moduleAlias.addAliases(aliases);
